added short data type fix

+++ src/dotnet/clrfunc.cpp
@@ -183,7 +183,7 @@ Handle<v8::Value> ClrFunc::MarshalCLRToV8(System::Object^ netdata)
     }
     else if (type == System::Int16::typeid)
     {
-        jsdata = Nan::New<v8::Integer>((int)netdata);
+        jsdata = Nan::New<v8::Integer>((short)netdata);
     }
     else if (type == System::Int64::typeid)
     {
